+++ src/Simulation/Motorbike.java
@@ -10,7 +10,7 @@ public class Motorbike extends Vehicle {
      * @param id             the vehicle id
      * @param vehicleBreadth the vehicle breadth - Half of its length
      */
-    public Motorbike(int speed, int positionOnRoad, int id, float vehicleBreadth) {
+    public Motorbike(int speed, int positionOnRoad, int id, int vehicleBreadth) {
         super(speed, positionOnRoad, id, vehicleBreadth);
     }
